requirements.txt
Babel==2.3.4
boto==2.45.0
dj-database-url==0.4.1
Django==1.10.2
django-appconf==1.0.2
django-bootstrap-form==3.2.1
django-ckeditor==5.1.1
django-constance==1.3.4
django-debug-toolbar==1.6
django-filter==0.15.3
django-flatblocks==0.9.3
django-imagekit==3.3
django-phonenumber-field==1.1.0
django-picklefield==0.3.2
django-rosetta==0.7.12
django-storages==1.5.1
django-suit==0.2.21
django-suit-ckeditor==0.0.2
djangorestframework==3.5.3
dropbox==7.1.1
gunicorn==19.6.0
Markdown==2.6.7
microsofttranslator==0.8
phonenumberslite==7.7.3
pilkit==1.1.13
Pillow==3.4.2
polib==1.0.8
psycopg2==2.6.2
pytz==2016.7
requests==2.12.4
six==1.10.0
sqlparse==0.2.2
typing==3.5.2.2
ujson==1.35
urllib3==1.19.1
whitenoise==3.2.2
